requirements.txt
absl-py==1.2.0
accelerate==0.12.0
aiohttp==3.8.1
aiosignal==1.2.0
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bitsandbytes==0.33.1
cachetools==5.2.0
charset-normalizer==2.1.1
clip @ git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1
datasets==2.4.0
decorator==5.1.1
diffusers @ git+https://github.com/huggingface/diffusers@6b02323a602a66841729c3a5d60844b24aa81ff2
dill==0.3.5.1
executing==1.0.0
filelock==3.8.0
frozenlist==1.3.1
fsspec==2022.8.2
ftfy==6.1.1
gdown==4.5.1
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
huggingface-hub==0.11.0
idna==3.3
importlib-metadata==4.12.0
ipdb==0.13.9
ipython==8.4.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
modelcards==0.1.6
multidict==6.0.2
multiprocess==0.70.13
numpy==1.23.2
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
regex==2022.8.17
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.5.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tokenizers==0.13.2
toml==0.10.2
torch==1.13.0+cu116
torchaudio==0.13.0+cu116
torchvision==0.14.0+cu116
tqdm==4.64.0
traitlets==5.3.0
transformers==4.24.0
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
xxhash==3.0.0
yarl==1.8.1
zipp==3.8.1
